Added more photos and retrained model. Want to try putting it together

+++ grabChamps.py
@@ -12,11 +12,11 @@ monitor_y = screeninfo.get_monitors()[0].height
 
 xRatio = monitor_x/2560
 yRatio = monitor_y/1440
-yTop = 1237 * yRatio
-yBottom = 1425 * yRatio
-xStart = 640 * xRatio
-xWidth = 257 * xRatio
-xSpacing = 268 * xRatio
+yTop = int(1237 * yRatio)
+yBottom = int(1425 * yRatio)
+xStart = int(640 * xRatio)
+xWidth = int(257 * xRatio)
+xSpacing = int(268 * xRatio)
 shopPath = os.path.join(os.getcwd(), "shopChamps")
 
 def screenGrabShop():
